requirements.txt
# DMA RAG System - Comprehensive Requirements
# Core ML and AI Libraries
sentence-transformers>=3.0.0
torch>=2.2.0
numpy>=1.24.0

# Vector Database
pinecone-client==5.0.1

# LLM Integration
google-generativeai>=0.7.0

# Web Framework (FastAPI for main server)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Alternative Web Framework (Flask for compatibility)
flask>=3.0.0
flask-cors>=4.0.0

# Web Scraping and Data Processing
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
lxml>=4.9.0

# Natural Language Processing
nltk>=3.8

# Progress Bars and Utilities
tqdm>=4.66.0

# Environment and Configuration
python-dotenv>=1.0.1

# Data Serialization
pydantic>=2.0.0

# HTTP Client
httpx>=0.25.0

# File Processing
openpyxl>=3.1.0

# Development and Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Additional utilities
python-dateutil>=2.8.0
pytz>=2023.3